Fixed multiple textarea autoresize in older IEs
Calling the supportsInputEvent method for the first time caused an
'oninput' property to appear on the body element, so the next calls
always returned true, even if the event is actually unsupported by
browser. This resulted in autoResize not working on all textareas other
than the first one which it was applied to. Also corrected the
'tabindex' attribute name.